+++ AddInManager/View/FrmAddInManager.xaml.cs
@@ -1,4 +1,5 @@
-﻿using RevitAddinManager.ViewModel;
+﻿using System.Diagnostics;
+using RevitAddinManager.ViewModel;
 using System.Windows;
 using System.Windows.Input;
 using RevitAddinManager.Model;
@@ -18,7 +19,8 @@ public partial class FrmAddInManager : Window
         this.LoadViewFromUri("/RevitAddinManager;component/view/frmaddinmanager.xaml");
         DataContext = vm;
         viewModel = vm;
-        vm.FrmAddInManager = this;
+        App.FrmAddInManager = this;
+        ThemManager.ChangeThem(true);
     }
 
     private void TbxDescription_OnLostKeyboardFocus(object sender, KeyboardFocusChangedEventArgs e)
@@ -62,6 +64,40 @@ public partial class FrmAddInManager : Window
 
     private void HandleTreeViewCommandKeyPress(object sender, KeyEventArgs e)
     {
+        int indexCmd = TreeViewCommand.Items.IndexOf(TreeViewCommand.SelectedItem);
+        if (e.Key == Key.Up && TabCommand.IsFocused)
+        {
+            tbxSearch.Focus();
+        }
+        else if (e.Key == Key.Up && indexCmd==0 && TabCommand.IsSelected)
+        {
+            TabCommand.Focus();
+        }
+        if (e.Key == Key.Down && TabCommand.IsSelected)
+        {
+            TreeViewCommand.Focus();
+        }
+        if (e.Key == Key.Enter)
+        {
+            viewModel.ExecuteAddinCommandClick();
+        }
+
+    }
+    private void HandleTreeViewAppKeyPress(object sender, KeyEventArgs e)
+    {
+        int indexCmd = TreeViewApp.Items.IndexOf(TreeViewApp.SelectedItem);
+        if (e.Key == Key.Up && TabApp.IsFocused)
+        {
+            tbxSearch.Focus();
+        }
+        else if (e.Key == Key.Up && indexCmd==0 && TabApp.IsSelected)
+        {
+            TabApp.Focus();
+        }
+        if (e.Key == Key.Down && TabApp.IsSelected)
+        {
+            TreeViewApp.Focus();
+        }
         if (e.Key == Key.Enter)
         {
             viewModel.ExecuteAddinCommandClick();
@@ -92,4 +128,10 @@ public partial class FrmAddInManager : Window
         }
         
     }
+
+    private void CloseFormEvent(object sender, KeyEventArgs e)
+    {
+        if (e.Key == Key.Escape) Close();
+    }
+
 }