+++ AddInManager/ViewModel/AddInManagerViewModel.cs
@@ -9,6 +9,7 @@ using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.IO;
 using System.Reflection;
+using System.Runtime.InteropServices.WindowsRuntime;
 using System.Text;
 using System.Windows;
 using System.Windows.Input;
@@ -21,7 +22,6 @@ public class AddInManagerViewModel : ViewModelBase
     private string Message { get; set; }
     private ElementSet Elements { get; set; }
     private RevitEvent RevitEvent = new RevitEvent();
-    public FrmAddInManager FrmAddInManager { get; set; }
     public AssemLoader AssemLoader { get; set; }
 
     public AddinManagerBase MAddinManagerBase { get; set; }
@@ -152,6 +152,7 @@ public class AddInManagerViewModel : ViewModelBase
     }
 
     public ICommand HelpCommand => new RelayCommand(HelpCommandClick);
+    public ICommand ChangeThemCommand => new RelayCommand(() => ThemManager.ChangeThem(false));
 
     private string vendorDescription = string.Empty;
 
@@ -189,6 +190,7 @@ public class AddInManagerViewModel : ViewModelBase
         set => OnPropertyChanged(ref isCanRun, value);
     }
 
+
     private bool isTabStartSelected;
 
     public bool IsTabStartSelected
@@ -200,6 +202,7 @@ public class AddInManagerViewModel : ViewModelBase
         }
         set => OnPropertyChanged(ref isTabStartSelected, value);
     }
+
     private bool isTabLogSelected;
 
     public bool IsTabLogSelected
@@ -208,11 +211,13 @@ public class AddInManagerViewModel : ViewModelBase
         {
             if (isTabLogSelected)
             {
-                LogControlViewModel vm = new LogControlViewModel() { FrmLogControl = FrmAddInManager.LogControl };
-                FrmAddInManager.LogControl.DataContext = vm;
-                FrmAddInManager.LogControl.Loaded += vm.LogFileWatcher;
-                FrmAddInManager.LogControl.Unloaded += vm.UserControl_Unloaded;
-            };
+                LogControlViewModel vm = new LogControlViewModel();
+                App.FrmAddInManager.LogControl.DataContext = vm;
+                App.FrmAddInManager.LogControl.Loaded += vm.LogFileWatcher;
+                App.FrmAddInManager.LogControl.Unloaded += vm.UserControl_Unloaded;
+            }
+
+            ;
             return isTabLogSelected;
         }
         set => OnPropertyChanged(ref isTabLogSelected, value);
@@ -222,7 +227,6 @@ public class AddInManagerViewModel : ViewModelBase
     {
         Process.Start("https://github.com/chuongmep/RevitAddInManager/wiki");
     }
-
     public AddInManagerViewModel(ExternalCommandData data, ref string message, ElementSet elements)
     {
         AssemLoader = new AssemLoader();
@@ -294,7 +298,7 @@ public class AddInManagerViewModel : ViewModelBase
                 CheckCountSelected(CommandItems, out var result);
                 if (result > 0)
                 {
-                    FrmAddInManager.Close();
+                    App.FrmAddInManager.Close();
                     RevitEvent.Run(Execute, false, null, null, false);
                 }
             }
@@ -445,12 +449,12 @@ public class AddInManagerViewModel : ViewModelBase
         {
             case AddinType.Command:
                 IsTabCmdSelected = true;
-                FrmAddInManager.TabCommand.Focus();
+                App.FrmAddInManager.TabCommand.Focus();
                 break;
 
             case AddinType.Application:
                 IsTabAppSelected = true;
-                FrmAddInManager.TabApp.Focus();
+                App.FrmAddInManager.TabApp.Focus();
                 break;
 
             case AddinType.Mixed:
@@ -594,9 +598,9 @@ public class AddInManagerViewModel : ViewModelBase
 
     private void ShowSuccessfully()
     {
-        MessageBox.Show(FrmAddInManager, "Save Successfully", Resource.AppName, MessageBoxButton.OK,
+        MessageBox.Show(App.FrmAddInManager, "Save Successfully", Resource.AppName, MessageBoxButton.OK,
             MessageBoxImage.Information);
-        FrmAddInManager.Close();
+        App.FrmAddInManager.Close();
     }
 
     private void FreshSearchClick()
@@ -652,7 +656,7 @@ public class AddInManagerViewModel : ViewModelBase
         var revitAddins = GetAddinFromFolder(Folder1);
         var addinsProgramData = GetAddinFromFolder(Folder2);
         var addinsPlugins = GetAddinFromFolder(Folder3);
-        revitAddins.ForEach(delegate (RevitAddin x)
+        revitAddins.ForEach(delegate(RevitAddin x)
         {
             addinStartup.Add(x);
             x.IsReadOnly = true;
@@ -706,12 +710,12 @@ public class AddInManagerViewModel : ViewModelBase
 
     private void SetToggleVisible()
     {
-        foreach (RevitAddin revitAddin in FrmAddInManager.DataGridStartup.SelectedItems)
+        foreach (RevitAddin revitAddin in App.FrmAddInManager.DataGridStartup.SelectedItems)
         {
             revitAddin.SetToggleState();
         }
 
-        FrmAddInManager.Close();
+        App.FrmAddInManager.Close();
         MessageBox.Show(Resource.Successfully, Resource.AppName);
     }
 
@@ -738,7 +742,7 @@ public class AddInManagerViewModel : ViewModelBase
                 var filePaths = Directory.GetFiles(folder).Where(x => x.Contains(DefaultSetting.FileName)).ToArray();
                 if (filePaths.Length == 0)
                 {
-                    MessageBox.Show(FrmAddInManager, "File Empty!", Resource.AppName, MessageBoxButton.OK,
+                    MessageBox.Show(App.FrmAddInManager, "File Empty!", Resource.AppName, MessageBoxButton.OK,
                         MessageBoxImage.Exclamation);
                     return;
                 }
@@ -759,7 +763,8 @@ public class AddInManagerViewModel : ViewModelBase
 
     private void EditAddinCommandClick()
     {
-        if (FrmAddInManager.DataGridStartup.SelectedItem is RevitAddin revitAddin && File.Exists(revitAddin.FilePath))
+        if (App.FrmAddInManager.DataGridStartup.SelectedItem is RevitAddin revitAddin &&
+            File.Exists(revitAddin.FilePath))
         {
             Process.Start(revitAddin.FilePath);
         }
@@ -771,7 +776,8 @@ public class AddInManagerViewModel : ViewModelBase
 
     private void OpenLocalAddinCommandClick()
     {
-        if (FrmAddInManager.DataGridStartup.SelectedItem is RevitAddin revitAddin && File.Exists(revitAddin.FilePath))
+        if (App.FrmAddInManager.DataGridStartup.SelectedItem is RevitAddin revitAddin &&
+            File.Exists(revitAddin.FilePath))
         {
             Process.Start("explorer.exe", "/select, " + revitAddin.FilePath);
         }
@@ -780,5 +786,4 @@ public class AddInManagerViewModel : ViewModelBase
             MessageBox.Show(Resource.FileNotFound, Resource.AppName);
         }
     }
-
 }